+++ CHANGELOG.md
@@ -170,6 +170,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#759](https://github.com/nix-rust/nix/pull/759))
 - Added `nix::unistd::{chown, lchown, fchown, fchownat}`
   ([#760](https://github.com/nix-rust/nix/pull/760))
+- Added `nix::stat::{futimens, utimensat}`
+  ([#761](https://github.com/nix-rust/nix/pull/761))
 
 ### Changed
 - Exposed the `mqueue` module for all supported operating systems.